+++ app/Http/Controllers/Controller.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Routing\Controller as BaseController;
+use App\Traits\ApiResponse;
 use OpenApi\Attributes as OA;
 
 /**
@@ -38,4 +39,5 @@ use OpenApi\Attributes as OA;
 abstract class Controller extends BaseController
 {
     use AuthorizesRequests;
+    use ApiResponse;
 }